Moved Client Constructor to it's own file. INIT now used for setting version, other uses TBD
